Refresh the WADL and test data
Both were a bit outdated, with missing fields here and there. I noticed
when working on some other fixes that necessitated to touch the fixup.xdl
file (still pending).

I don't know whether the change Vernooĳ->Vernooij was due to a user
change or some change in the backend, but it's the data I got back from
LP directly.