Fix the types for distributions items
Some items were wrongly overridden to bool, some weren't marked as
options.

This also moves the bool override for various distribution-only fields
from build.rs to the xsl sheet, as the build step of the library is
unbelievably slow and I'm guessing the massive match on every property
name could be one of the culprits. I haven't measured, though.